Move pformat_json into utils
